test : LottoIssuerTest의 Lottos 생성 테스트 LottosTest로 이관
